Cambios BD y guardado de ordenes
Se agregaron archivos de repositorio y servicio, así como algunos cambios en las entidades, para permitir el guardado de las ordenes y los detalles de las mismas en la base de datos. Ahora se permite la suma de cantidades de los productos al agregarlos al carrito.  Se agregan algunos cambios en la base de datos con respecto a las ordenes y detalles, así como a un valor del menú

+++ src/main/java/com/restaurante/entity/DetalleOrden.java
@@ -14,7 +14,7 @@ public class DetalleOrden implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nombre;
     private double cantidad;
     private double precio;
@@ -30,7 +30,7 @@ public class DetalleOrden implements Serializable {
 
     }
 
-    public DetalleOrden(Integer id, String nombre, double cantidad, double precio, double total) {
+    public DetalleOrden(Long id, String nombre, double cantidad, double precio, double total) {
         super();
         this.id = id;
         this.nombre = nombre;
@@ -39,11 +39,11 @@ public class DetalleOrden implements Serializable {
         this.total = total;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
